--- a/mainchain/tracers/native/noop.go
+++ b/mainchain/tracers/native/noop.go
@@ -29,7 +29,7 @@ import (
 )
 
 func init() {
-	register("noopTracerNative", newNoopTracer)
+	register("noopTracer", newNoopTracer)
 }
 
 // noopTracer is a go implementation of the Tracer interface which