--- a/mainchain/tracers/native/replay.go
+++ b/mainchain/tracers/native/replay.go
@@ -1,0 +1,34 @@
+/*
+ *  Copyright 2022 KardiaChain
+ *  This file is part of the go-kardia library.
+ *
+ *  The go-kardia library is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Lesser General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  The go-kardia library is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU Lesser General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Lesser General Public License
+ *  along with the go-kardia library. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package native
+
+import (
+	"github.com/kardiachain/go-kardia/mainchain/tracers"
+)
+
+func init() {
+	register("replayTracer", newReplayTracer)
+}
+
+func newReplayTracer() tracers.Tracer {
+	return &callTracer{
+		callstack: make([]callFrame, 1),
+		name:      "replayTracer",
+	}
+}